Add context managers and move MultiFile to own module
* Add __enter__, __exit__ and close methods to MultiFile
* Call MultiFile.close in LSTEventSource.__exit__ and
  LSTEventSource.close
